update transactions handler
